+++ ecommerce/src/Components/Administrator/ShopOrderComponent/ShopOrderLine.jsx
@@ -14,44 +14,54 @@ const ShopOrderLine = (props) => {
     const context = useContext(OrdersContext);
     const orders = context.orders;
     const updateOrders = context.setOrders;
+    const setNotify = context.setNotify;
+    const setConfirmDialog = context.setConfirmDialog;
 
     const navigate = useNavigate();
-    useEffect(()=>{
+    useEffect(() => {
         setOrders(props.order);
-    },[props])
+    }, [props])
 
-    function processDate(dateTime){
+    function processDate(dateTime) {
         let MonthsOfYear = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"];
         let date = new Date(dateTime);
-        let result = date.getDate() +" "+ MonthsOfYear[date.getMonth()] + " " + date.getFullYear();
+        let result = date.getDate() + " " + MonthsOfYear[date.getMonth()] + " " + date.getFullYear();
         return result;
     }
 
-    function changeOrderStatus(event){
+    function changeOrderStatus(event) {
         let select = event.currentTarget.value;
-        let status = props.status.filter(st=> parseInt(st.id) === parseInt(select));
-        setOrders({...order, orderStatus: status[0]});
+        let status = props.status.filter(st => parseInt(st.id) === parseInt(select));
+        setOrders({ ...order, orderStatus: status[0] });
     }
 
-    function quickUpdateOrderStatus(){
-        ShopOrderService.updateShopOrderStatus(order.id, order.orderStatus.id).then(response=>{
+    function quickUpdateOrderStatus() {
+        ShopOrderService.updateShopOrderStatus(order.id, order.orderStatus.id).then(response => {
             console.log(response.data);
+            setNotify({ isOpen: true, message: "Update successful!", type: "success" })
         });
     }
 
-    function goToDetail(){
-        navigate(order.id+"/detail");
+    function goToDetail() {
+        navigate(order.id + "/detail");
     }
 
-    function deleteShopOder(){
-        ShopOrderService.deleteShopOrder(order.id).then(response=>{
-            console.log("Delete Result: ", response.data);
-        })
-        let ls = orders;
-        let el = ls.filter(or => or.id === order.id);
-        let index = ls.indexOf(el[0]);
-        ls.splice(index, 1);
-        updateOrders([...ls]);
+    function deleteShopOder() {
+        setConfirmDialog({
+            isOpen: true,
+            title: "Are you sure delete this record?",
+            subTitle: "You can't undo this operation.",
+            commit: () => {
+                ShopOrderService.deleteShopOrder(order.id).then(response => {
+                    console.log("Delete Result: ", response.data);
+                })
+                let ls = orders;
+                let el = ls.filter(or => or.id === order.id);
+                let index = ls.indexOf(el[0]);
+                ls.splice(index, 1);
+                updateOrders([...ls]);
+            }
+        });
     }
 
     return (
@@ -60,7 +70,7 @@ const ShopOrderLine = (props) => {
             <td style={{ textAlign: "center" }}>{processDate(order.dateCreate)}</td>
             <td style={{ textAlign: "center" }}>{order.siteUser.name}</td>
             <td style={{ textAlign: "center" }}>
-                <select className={'status-color-'+order.orderStatus.id} id={"order-status-" + order.id} value={order.orderStatus.id} onChange={(event) => changeOrderStatus(event)}>
+                <select className={'status-color-' + order.orderStatus.id} id={"order-status-" + order.id} value={order.orderStatus.id} onChange={(event) => changeOrderStatus(event)}>
                     {
                         props.status.map((st, index) =>
                             <option key={index} value={st.id}>{st.status}</option>
@@ -68,16 +78,16 @@ const ShopOrderLine = (props) => {
                     }
                 </select>
             </td>
-            <td>
+            <td className='price-color'>
                 {Intl.NumberFormat('en-US', { style: "currency", currency: "USD" }).format(order.orderTotal)}
             </td>
             <td style={{ textAlign: "center" }} className='m-1'>
                 <div>
-                    <button className='btn text-success' title='Confirm' onClick={quickUpdateOrderStatus}><CloudUploadIcon/></button>
-                    <button className='btn text-light' title='Info' onClick={goToDetail}><InfoIcon/></button>
-                    <button className='btn text-danger' title='Delete' onClick={deleteShopOder}><DeleteIcon/></button>
+                    <button className='btn text-success' title='Confirm' onClick={quickUpdateOrderStatus}><CloudUploadIcon /></button>
+                    <button className='btn' style={{ color: "#0d6efd" }} title='Info' onClick={goToDetail}><InfoIcon /></button>
+                    <button className='btn text-danger' title='Delete' onClick={deleteShopOder}><DeleteIcon /></button>
                 </div>
-                
+
             </td>
         </tr>
     );
